+++ CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2.1.1-pre.1
+## [Unreleased]
 
+### Added
+
+### Changed
+
+## [2.1.1] - 2023-11-15
 ### Added
 - console message counters and last log message in bot of the dashboard